+++ src/components/service-navigation-language-select/jsdom.test.js
@@ -4,19 +4,19 @@ import axe from '../../../lib/axe-helper';
 
 import { render, getExamples } from '../../../lib/jest-helpers';
 
-const examples = getExamples('language-select');
+const examples = getExamples('service-navigation-language-select');
 
 describe('New Tab Link', () => {
   describe('by default', () => {
     it('passes accessibility tests', async () => {
-      const $ = render('language-select', examples.default);
+      const $ = render('service-navigation-language-select', examples.default);
 
       const results = await axe($.html());
       expect(results).toHaveNoViolations();
     });
 
     it('renders an English element as text and Welsh as a link', () => {
-      const $ = render('language-select', examples.default);
+      const $ = render('service-navigation-language-select', examples.default);
 
       const $component = $('.govuk-link');
       expect($component.length).toEqual(1);
@@ -28,14 +28,14 @@ describe('New Tab Link', () => {
       expect($component.data('journeyClick')).toEqual('link - click:lang-select:Cymraeg');
 
       expect($component.find('.govuk-visually-hidden').eq(0).text()).toEqual('Newid yr iaith ir Gymraeg');
-      expect($component.find('[aria-hidden="true"]').eq(0).text()).toEqual('Cymraeg');
-      expect($.text().trim().replace(/[\s,\n]+/g, ' ')).toEqual('English Newid yr iaith ir Gymraeg Cymraeg');
+      expect($component.find('[aria-hidden="true"]').eq(0).text()).toEqual('CYM');
+      expect($.text().trim().replace(/[\s,\n]+/g, ' ')).toEqual('ENG Newid yr iaith ir Gymraeg CYM');
 
-      expect($('.hmrc-language-select__list-item').find('[aria-current="true"]').eq(0).text()).toEqual('English');
+      expect($('.hmrc-service-navigation-language-select__list-item').find('[aria-current="true"]').eq(0).text()).toEqual('ENG');
     });
 
     it('renders a Welsh element as text and English as a link', () => {
-      const $ = render('language-select', examples.welsh);
+      const $ = render('service-navigation-language-select', examples.welsh);
 
       const $component = $('.govuk-link');
       expect($component.length).toEqual(1);
@@ -47,10 +47,10 @@ describe('New Tab Link', () => {
       expect($component.data('journeyClick')).toEqual('link - click:lang-select:English');
 
       expect($component.find('.govuk-visually-hidden').eq(0).text()).toEqual('Change the language to English');
-      expect($component.find('[aria-hidden="true"]').eq(0).text()).toEqual('English');
-      expect($.text().trim().replace(/[\s,\n]+/g, ' ')).toEqual('Change the language to English English Cymraeg');
+      expect($component.find('[aria-hidden="true"]').eq(0).text()).toEqual('ENG');
+      expect($.text().trim().replace(/[\s,\n]+/g, ' ')).toEqual('Change the language to English ENG CYM');
 
-      expect($('.hmrc-language-select__list-item').find('[aria-current="true"]').eq(0).text()).toEqual('Cymraeg');
+      expect($('.hmrc-service-navigation-language-select__list-item').find('[aria-current="true"]').eq(0).text()).toEqual('CYM');
     });
   });
 });